# search: BEAM_UOPS_MAX to prune candidates with too many uops