not the highest possible performance, but adequate for most usage

TODO: will add testing to show performance before and after