@flammit 2024-04-06
Test commands:
```sh
# LLAMA-7B BEAM=2
time PYTHONPATH=. JIT=1 BEAM=2 CACHELEVEL=0 python3 examples/llama.py --gen 1 --prompt "Hello." --count 10 --temperature 0 --timing

# GPT2 BEAM=2
time PYTHONPATH=. JIT=1 HALF=1 BEAM=2 CACHELEVEL=0 CAST_BEFORE_VIEW=0 python3 examples/gpt2.py --count 10 --temperature 0 --timing

# CIFAR LATEBEAM=1
time PYTHONPATH=. STEPS=10 DEFAULT_FLOAT=HALF LATEBEAM=1 CACHELEVEL=0 python3 examples/hlb_cifar10.py

# CIFAR LATEBEAM=8
time PYTHONPATH=. STEPS=10 DEFAULT_FLOAT=HALF LATEBEAM=8 CACHELEVEL=0 python3 examples/hlb_cifar10.py
```

| Backend | Command | Master Total Time | Master Step Time | Branch Total Time | Branch Step Time |
| --- | --- | --- | --- | --- | --- |
| METAL | LLAMA-7B BEAM=2       | 37s | 45ms | 28s | 46ms |
| METAL | GPT2 FP16 BEAM=2      | 30s |  5ms | 16s |  5ms |
| METAL | CIFAR FP16 LATEBEAM=1 before TC fix | 5m43s | 141ms | N/A | N/A |
| METAL | CIFAR FP16 LATEBEAM=1 | 6m11s | 105-106ms | 5m06s | 114ms (!) |
| METAL | CIFAR FP16 LATEBEAM=1 BEAM_MIN_PROGRESS=2 | N/A | N/A | 5m03s | 111-112ms (!) |
| METAL | CIFAR FP16 LATEBEAM=1 BEAM_UPCAST_MAX=256 | N/A | N/A | 6m12s | 100-102ms |
| METAL | CIFAR FP16 LATEBEAM=8 | 54m | 97ms | 26m | 110ms (!) |
| METAL | CIFAR FP16 LATEBEAM=8 BEAM_UPCAST_MAX=256 | N/A | N/A | 33m | 98ms |
| HSA   | LLAMA-7B BEAM=2       | 66s | 22ms | 52s | 22ms |
| HSA   | GPT2 FP16 BEAM=2      | 67s | 3.7ms | 51s | 3.7ms |
| HSA   | CIFAR FP16 LATEBEAM=1 | 20m | 35-40ms | 6m40s | 36-40ms |
| HSA   | CIFAR FP16 LATEBEAM=1 BEAM_MIN_PROGRESS=2 | N/A | N/A | 6m53s | 39-41ms |
| HSA   | CIFAR FP16 LATEBEAM=1 BEAM_UPCAST_MAX=256 | N/A | N/A | 8m01s | 39-41ms |
| HSA   | CIFAR FP16 LATEBEAM=8 | 43m | 33ms | 18m | 30ms |
| HSA   | CIFAR FP16 LATEBEAM=8 BEAM_UPCAST_MAX=256 | N/A | N/A | 24m | 37ms |
| CUDA  | LLAMA-7B BEAM=2       | 70-80s | 19ms | 72s | 19ms |
| CUDA  | GPT2 FP16 BEAM=2      | 70s | 1.7ms | 58s | 1.7ms |
| CUDA  | CIFAR FP16 LATEBEAM=1 | 6m | 36-38ms | 6m | 33-37ms |
| CUDA  | CIFAR FP16 LATEBEAM=1 BEAM_MIN_PROGRESS=2 | N/A | N/A | 6m25s | 33-34ms |
| CUDA  | CIFAR FP16 LATEBEAM=1 BEAM_UPCAST_MAX=256 | N/A | N/A | 6m00s | 30-32ms |
| CUDA  | CIFAR FP16 LATEBEAM=8 | 33m | 29ms | 33m | 30ms |
| CUDA  | CIFAR FP16 LATEBEAM=8 BEAM_UPCAST_MAX=256 | N/A | N/A | OOM (not tiny only 32gbs) | OOM |