@flammit 2024-04-06
Added one more attempt at CIFAR: `CUDA=1 LATEBEAM=8 BEAM_UPCAST_MAX=256 BEAM_MAX_TASKS_PER_CHILD=1` but it didn't finish after more than 17hrs.