@chenyuxyz 2024-04-06
maybe a version of `BEAM_MIN_PROGRESS` that relies on relative time can mitigate the slowdown issue.