@flammit 2024-04-06
My mistake - it wasn't running CUDA=1... my drivers didn't load after reboot after reinstalling and rerunning, it finished in 43 minutes.