@chenyuxyz 2024-04-09
can we get a version with `BEAM_MAX_TASKS_PER_CHILD` change and uops MAX only first? i think these 2 are the least controversial