@flammit 2024-04-09
Done!  also added sanity line to .gitignore as discussed