@flammit 2024-04-15
> how did you pick 3000 as the default threshold, are we hitting it currently for the usual examples?

it was just a heuristic based on looking at `BEAM_LOG=1` outputs as a proxy for 1-ish second compile times (here's one from a tiny red): 

<img width="1210" alt="Screenshot 2024-04-15 at 1 33 38 PM" src="https://github.com/tinygrad/tinygrad/assets/1693995/dd06d837-758f-4472-b188-c8a7251eb370">