@chenyuxyz 2024-04-15
will measure resnet compile time again after this change